cue/load: move code before refactoring
This step prepares for duplicating much of the
loader logic so that it can be modified for module
loading.

This CL only moves code, with no changes at
all other than just the code movement, so
the next CL will make any changes more clear.
