+++ cue/load/import.go
@@ -15,40 +15,17 @@
 package load
 
 import (
-	"bytes"
+	"fmt"
 	"os"
+	pathpkg "path"
 	"path/filepath"
 	"sort"
-	"strconv"
 	"strings"
-	"unicode"
-	"unicode/utf8"
 
-	"cuelang.org/go/cue/ast"
 	"cuelang.org/go/cue/build"
 	"cuelang.org/go/cue/errors"
-	"cuelang.org/go/cue/parser"
 	"cuelang.org/go/cue/token"
-	"cuelang.org/go/internal"
 	"cuelang.org/go/internal/filetypes"
-)
-
-// An importMode controls the behavior of the Import method.
-type importMode uint
-
-const (
-	// If findOnly is set, Import stops after locating the directory
-	// that should contain the sources for a package. It does not
-	// read any files in the directory.
-	findOnly importMode = 1 << iota
-
-	// If importComment is set, parse import comments on package statements.
-	// Import returns an error if it finds a comment it cannot understand
-	// or finds conflicting comments in multiple source files.
-	// See golang.org/s/go14customimport for more information.
-	importComment
-
-	allowAnonymous
 )
 
 // importPkg returns details about the CUE package named by the import path,
@@ -260,364 +237,106 @@ func (l *loader) loadFunc() build.LoadFunc {
 	}
 }
 
-func rewriteFiles(p *build.Instance, root string, isLocal bool) {
-	p.Root = root
+// newRelInstance returns a build instance from the given
+// relative import path.
+func (c *Config) newRelInstance(pos token.Pos, path, pkgName string) *build.Instance {
+	if !isLocalImport(path) {
+		panic(fmt.Errorf("non-relative import path %q passed to newRelInstance", path))
+	}
+	fs := c.fileSystem
 
-	normalizeFiles(p.BuildFiles)
-	normalizeFiles(p.IgnoredFiles)
-	normalizeFiles(p.OrphanedFiles)
-	normalizeFiles(p.InvalidFiles)
-	normalizeFiles(p.UnknownFiles)
+	var err errors.Error
+	dir := path
+
+	p := c.Context.NewInstance(path, c.loadFunc)
+	p.PkgName = pkgName
+	p.DisplayPath = filepath.ToSlash(path)
+	// p.ImportPath = string(dir) // compute unique ID.
+	p.Root = c.ModuleRoot
+	p.Module = c.Module
+
+	dir = filepath.Join(c.Dir, filepath.FromSlash(path))
+
+	if path != cleanImport(path) {
+		err = errors.Append(err, c.loader.errPkgf(nil,
+			"non-canonical import path: %q should be %q", path, pathpkg.Clean(path)))
+	}
+
+	if importPath, e := c.importPathFromAbsDir(fsPath(dir), path); e != nil {
+		// Detect later to keep error messages consistent.
+	} else {
+		p.ImportPath = string(importPath)
+	}
+
+	p.Dir = dir
+
+	if fs.isAbsPath(path) || strings.HasPrefix(path, "/") {
+		err = errors.Append(err, errors.Newf(pos,
+			"absolute import path %q not allowed", path))
+	}
+	if err != nil {
+		p.Err = errors.Append(p.Err, err)
+		p.Incomplete = true
+	}
+
+	return p
 }
 
-// normalizeFiles sorts the files so that files contained by a parent directory
-// always come before files contained in sub-directories, and that filenames in
-// the same directory are sorted lexically byte-wise, like Go's `<` operator.
-func normalizeFiles(a []*build.File) {
-	sort.Slice(a, func(i, j int) bool {
-		fi := a[i].Filename
-		fj := a[j].Filename
-		ci := strings.Count(fi, string(filepath.Separator))
-		cj := strings.Count(fj, string(filepath.Separator))
-		if ci != cj {
-			return ci < cj
-		}
-		return fi < fj
-	})
-}
-
-type fileProcessor struct {
-	firstFile        string
-	firstCommentFile string
-	imported         map[string][]token.Pos
-	allTags          map[string]bool
-	allFiles         bool
-	ignoreOther      bool // ignore files from other packages
-	allPackages      bool
-
-	c    *Config
-	pkgs map[string]*build.Instance
-	pkg  *build.Instance
-
-	err errors.Error
-}
-
-func newFileProcessor(c *Config, p *build.Instance) *fileProcessor {
-	return &fileProcessor{
-		imported: make(map[string][]token.Pos),
-		allTags:  make(map[string]bool),
-		c:        c,
-		pkgs:     map[string]*build.Instance{"_": p},
-		pkg:      p,
-	}
-}
-
-func countCUEFiles(c *Config, p *build.Instance) int {
-	count := len(p.BuildFiles)
-	for _, f := range p.IgnoredFiles {
-		if c.Tools && strings.HasSuffix(f.Filename, "_tool.cue") {
-			count++
-		}
-		if c.Tests && strings.HasSuffix(f.Filename, "_test.cue") {
-			count++
-		}
-	}
-	return count
-}
-
-func (fp *fileProcessor) finalize(p *build.Instance) errors.Error {
-	if fp.err != nil {
-		return fp.err
-	}
-	if countCUEFiles(fp.c, p) == 0 &&
-		!fp.c.DataFiles &&
-		(p.PkgName != "_" || !fp.allPackages) {
-		fp.err = errors.Append(fp.err, &NoFilesError{Package: p, ignored: len(p.IgnoredFiles) > 0})
-		return fp.err
+func (c *Config) importPathFromAbsDir(absDir fsPath, key string) (importPath, errors.Error) {
+	if c.ModuleRoot == "" {
+		return "", errors.Newf(token.NoPos,
+			"cannot determine import path for %q (root undefined)", key)
 	}
 
-	for tag := range fp.allTags {
-		p.AllTags = append(p.AllTags, tag)
-	}
-	sort.Strings(p.AllTags)
-
-	p.ImportPaths, _ = cleanImports(fp.imported)
-
-	return nil
-}
-
-func (fp *fileProcessor) add(pos token.Pos, root string, file *build.File, mode importMode) (added bool) {
-	fullPath := file.Filename
-	if fullPath != "-" {
-		if !filepath.IsAbs(fullPath) {
-			fullPath = filepath.Join(root, fullPath)
-		}
-	}
-	file.Filename = fullPath
-
-	base := filepath.Base(fullPath)
-
-	// special * and _
-	p := fp.pkg // default package
-
-	// badFile := func(p *build.Instance, err errors.Error) bool {
-	badFile := func(err errors.Error) bool {
-		fp.err = errors.Append(fp.err, err)
-		file.ExcludeReason = fp.err
-		p.InvalidFiles = append(p.InvalidFiles, file)
-		return true
+	dir := filepath.Clean(string(absDir))
+	if !strings.HasPrefix(dir, c.ModuleRoot) {
+		return "", errors.Newf(token.NoPos,
+			"cannot determine import path for %q (dir outside of root)", key)
 	}
 
-	match, data, err := matchFile(fp.c, file, true, fp.allFiles, fp.allTags)
+	pkg := filepath.ToSlash(dir[len(c.ModuleRoot):])
 	switch {
-	case match:
+	case strings.HasPrefix(pkg, "/cue.mod/"):
+		pkg = pkg[len("/cue.mod/"):]
+		if pkg == "" {
+			return "", errors.Newf(token.NoPos,
+				"invalid package %q (root of %s)", key, modDir)
+		}
 
-	case err == nil:
-		// Not a CUE file.
-		p.OrphanedFiles = append(p.OrphanedFiles, file)
-		return false
+		// TODO(legacy): remove.
+	case strings.HasPrefix(pkg, "/pkg/"):
+		pkg = pkg[len("/pkg/"):]
+		if pkg == "" {
+			return "", errors.Newf(token.NoPos,
+				"invalid package %q (root of %s)", key, pkgDir)
+		}
 
-	case !errors.Is(err, errExclude):
-		return badFile(err)
-
+	case c.Module == "":
+		return "", errors.Newf(token.NoPos,
+			"cannot determine import path for %q (no module)", key)
 	default:
-		file.ExcludeReason = err
-		if file.Interpretation == "" {
-			p.IgnoredFiles = append(p.IgnoredFiles, file)
-		} else {
-			p.OrphanedFiles = append(p.OrphanedFiles, file)
-		}
-		return false
+		pkg = c.Module + pkg
 	}
 
-	pf, perr := parser.ParseFile(fullPath, data, parser.ImportsOnly, parser.ParseComments)
-	if perr != nil {
-		badFile(errors.Promote(perr, "add failed"))
-		return true
+	name := c.Package
+	switch name {
+	case "_", "*":
+		name = ""
 	}
 
-	_, pkg, pos := internal.PackageInfo(pf)
-	if pkg == "" {
-		pkg = "_"
-	}
-
-	switch {
-	case pkg == p.PkgName, mode&allowAnonymous != 0:
-	case fp.allPackages && pkg != "_":
-		q := fp.pkgs[pkg]
-		if q == nil {
-			q = &build.Instance{
-				PkgName: pkg,
-
-				Dir:         p.Dir,
-				DisplayPath: p.DisplayPath,
-				ImportPath:  p.ImportPath + ":" + pkg,
-				Root:        p.Root,
-				Module:      p.Module,
-			}
-			fp.pkgs[pkg] = q
-		}
-		p = q
-
-	case pkg != "_":
-
-	default:
-		file.ExcludeReason = excludeError{errors.Newf(pos, "no package name")}
-		p.IgnoredFiles = append(p.IgnoredFiles, file)
-		return false // don't mark as added
-	}
-
-	if !fp.c.AllCUEFiles {
-		if err := shouldBuildFile(pf, fp); err != nil {
-			if !errors.Is(err, errExclude) {
-				fp.err = errors.Append(fp.err, err)
-			}
-			file.ExcludeReason = err
-			p.IgnoredFiles = append(p.IgnoredFiles, file)
-			return false
-		}
-	}
-
-	if pkg != "" && pkg != "_" {
-		if p.PkgName == "" {
-			p.PkgName = pkg
-			fp.firstFile = base
-		} else if pkg != p.PkgName {
-			if fp.ignoreOther {
-				file.ExcludeReason = excludeError{errors.Newf(pos,
-					"package is %s, want %s", pkg, p.PkgName)}
-				p.IgnoredFiles = append(p.IgnoredFiles, file)
-				return false
-			}
-			return badFile(&MultiplePackageError{
-				Dir:      p.Dir,
-				Packages: []string{p.PkgName, pkg},
-				Files:    []string{fp.firstFile, base},
-			})
-		}
-	}
-
-	isTest := strings.HasSuffix(base, "_test"+cueSuffix)
-	isTool := strings.HasSuffix(base, "_tool"+cueSuffix)
-
-	if mode&importComment != 0 {
-		qcom, line := findimportComment(data)
-		if line != 0 {
-			com, err := strconv.Unquote(qcom)
-			if err != nil {
-				badFile(errors.Newf(pos, "%s:%d: cannot parse import comment", fullPath, line))
-			} else if p.ImportComment == "" {
-				p.ImportComment = com
-				fp.firstCommentFile = base
-			} else if p.ImportComment != com {
-				badFile(errors.Newf(pos, "found import comments %q (%s) and %q (%s) in %s", p.ImportComment, fp.firstCommentFile, com, base, p.Dir))
-			}
-		}
-	}
-
-	for _, decl := range pf.Decls {
-		d, ok := decl.(*ast.ImportDecl)
-		if !ok {
-			continue
-		}
-		for _, spec := range d.Specs {
-			quoted := spec.Path.Value
-			path, err := strconv.Unquote(quoted)
-			if err != nil {
-				badFile(errors.Newf(
-					spec.Path.Pos(),
-					"%s: parser returned invalid quoted string: <%s>", fullPath, quoted,
-				))
-			}
-			if !isTest || fp.c.Tests {
-				fp.imported[path] = append(fp.imported[path], spec.Pos())
-			}
-		}
-	}
-	switch {
-	case isTest:
-		if fp.c.loader.cfg.Tests {
-			p.BuildFiles = append(p.BuildFiles, file)
-		} else {
-			file.ExcludeReason = excludeError{errors.Newf(pos,
-				"_test.cue files excluded in non-test mode")}
-			p.IgnoredFiles = append(p.IgnoredFiles, file)
-		}
-	case isTool:
-		if fp.c.loader.cfg.Tools {
-			p.BuildFiles = append(p.BuildFiles, file)
-		} else {
-			file.ExcludeReason = excludeError{errors.Newf(pos,
-				"_tool.cue files excluded in non-cmd mode")}
-			p.IgnoredFiles = append(p.IgnoredFiles, file)
-		}
-	default:
-		p.BuildFiles = append(p.BuildFiles, file)
-	}
-	return true
+	return addImportQualifier(importPath(pkg), name)
 }
 
-func findimportComment(data []byte) (s string, line int) {
-	// expect keyword package
-	word, data := parseWord(data)
-	if string(word) != "package" {
-		return "", 0
-	}
+func (c *Config) newInstance(pos token.Pos, p importPath) *build.Instance {
+	dir, name, err := c.absDirFromImportPath(pos, p)
+	i := c.Context.NewInstance(dir, c.loadFunc)
+	i.Dir = dir
+	i.PkgName = name
+	i.DisplayPath = string(p)
+	i.ImportPath = string(p)
+	i.Root = c.ModuleRoot
+	i.Module = c.Module
+	i.Err = errors.Append(i.Err, err)
 
-	// expect package name
-	_, data = parseWord(data)
-
-	// now ready for import comment, a // comment
-	// beginning and ending on the current line.
-	for len(data) > 0 && (data[0] == ' ' || data[0] == '\t' || data[0] == '\r') {
-		data = data[1:]
-	}
-
-	var comment []byte
-	switch {
-	case bytes.HasPrefix(data, slashSlash):
-		i := bytes.Index(data, newline)
-		if i < 0 {
-			i = len(data)
-		}
-		comment = data[2:i]
-	}
-	comment = bytes.TrimSpace(comment)
-
-	// split comment into `import`, `"pkg"`
-	word, arg := parseWord(comment)
-	if string(word) != "import" {
-		return "", 0
-	}
-
-	line = 1 + bytes.Count(data[:cap(data)-cap(arg)], newline)
-	return strings.TrimSpace(string(arg)), line
-}
-
-var (
-	slashSlash = []byte("//")
-	newline    = []byte("\n")
-)
-
-// skipSpaceOrComment returns data with any leading spaces or comments removed.
-func skipSpaceOrComment(data []byte) []byte {
-	for len(data) > 0 {
-		switch data[0] {
-		case ' ', '\t', '\r', '\n':
-			data = data[1:]
-			continue
-		case '/':
-			if bytes.HasPrefix(data, slashSlash) {
-				i := bytes.Index(data, newline)
-				if i < 0 {
-					return nil
-				}
-				data = data[i+1:]
-				continue
-			}
-		}
-		break
-	}
-	return data
-}
-
-// parseWord skips any leading spaces or comments in data
-// and then parses the beginning of data as an identifier or keyword,
-// returning that word and what remains after the word.
-func parseWord(data []byte) (word, rest []byte) {
-	data = skipSpaceOrComment(data)
-
-	// Parse past leading word characters.
-	rest = data
-	for {
-		r, size := utf8.DecodeRune(rest)
-		if unicode.IsLetter(r) || '0' <= r && r <= '9' || r == '_' {
-			rest = rest[size:]
-			continue
-		}
-		break
-	}
-
-	word = data[:len(data)-len(rest)]
-	if len(word) == 0 {
-		return nil, nil
-	}
-
-	return word, rest
-}
-
-func cleanImports(m map[string][]token.Pos) ([]string, map[string][]token.Pos) {
-	all := make([]string, 0, len(m))
-	for path := range m {
-		all = append(all, path)
-	}
-	sort.Strings(all)
-	return all, m
-}
-
-// isLocalImport reports whether the import path is
-// a local import path, like ".", "..", "./foo", or "../foo".
-func isLocalImport(path string) bool {
-	return path == "." || path == ".." ||
-		strings.HasPrefix(path, "./") || strings.HasPrefix(path, "../")
+	return i
 }
